Bump the ably-cocoa dependency to latest, version 1.2.9.
I tried to follow the steps that had worked for me a few days ago, as detailed in the opening comment here:
https://github.com/ably/ably-flutter/pull/293

However, I got some odd CocoaPods errors. I ended up following these instructions (it's `trunk`) which got me there for the example directory but was still failing for test_integration:
https://stackoverflow.com/a/34526412/392847

So, I ended up picking the changes that the pod tool had applied to the lock file in the example directory and manually transferring those over to the test_integration lock file.
Not ideal - so there's more to be understood here as this is not sustainable.