Make client handle reservation safe for use from any thread.
The existing reserveClientHandle() was assumed to always be called from the same thread, with no logical link to the createRest() and createRealtime() that were being called after it.
I'm honestly not sure if this was a bug that was being experienced right now, but the assumptions that the existing AblyMethodCallHandler implementation were making were likely to become an issue for us later on as the codebase grows, so I wanted to get this fixed now.