Make all AblyInstanceStore methods safe to be called from any thread.
I considered have a discrete lock for paginated results but kept on circling back to the implications of the work done by the reset() method.
So, on balance, I can't envisage a scenario where any of these methods take such a long time with the lock obtained that this change would have a measurable impact on performance.