Handle docstrings for constructors/properties that are not documented in canonical api ref in CipherParams class
Implementation for the whole `CipherParams` feature differs heavily from the canonical api ref, due to having to handle both iOS and Android implementations. Due to that I've decided to slightly edit the legacy docstrings and leave them in as "canonical".

Also, the whole CipherParamsInternal is marked as nodoc, since it's not exposed outside of the package.