Read authUrl as NSURL
authUrl was being read as string which wasn't treated as an error as type mismatch is allowed in objective C. Also macros do not seem to be checked by XCode. That's why there was also no warning about mismatch. I created yet another macro to read authURL as NSURL. I did this to prevent spaghetti of conventions.