fix: out of sync message handlers
Makes `StreamsChannel#_lastId` field static, because:

During hot reloading, the Ably instance can be recreated. `StreamsChannel#_lastId` field is used to create IDs for platform-specific subscriptions, which are used on iOS and Android. Code that used to register subscription handler is a singleton, so two Ably instances overwrite each other, causing the platform implementation to be out of sync with the Flutter message handler.